2021/2021-07-12/dean/02-api/requirements.txt
-e .
aws_cdk.aws_lambda
aws_cdk.aws_apigateway